Add fixes from review
